Moved parse_action() form .functions to .private
